==> Notebooks/2. data preparation.ipynb <==
import pandas as pd
import json

file_path= r'G:\NLP Projects\MitRA aTTACK\Notre projet\Data\tactic.json'
with open(file_path, 'r') as file:
    data = json.load(file)

data[1]

records = []

for tactic in data:
    tactic_name = tactic.get('tactic-name')
    tactic_id = tactic.get('ID')
    tactic_description = tactic.get('tactic')
    
    for technique_group in tactic.get('technique', []):
        for technique in technique_group:
            technique_name = technique.get('technique-name')
            technique_id = technique.get('technique-ID')
            technique_description = technique.get('technique-description')
            
            for sub_technique_group in technique.get('sub-technique', []):
                for sub_technique in sub_technique_group:
                    sub_technique_name = sub_technique.get('sub-technique-name')
                    sub_technique_number = sub_technique.get('number')
                    sub_technique_description = sub_technique.get('sub-technique-description')
                    
                    records.append({
                        'technique_id': technique_id,
                        'sub_technique_name': sub_technique_name,
                        'sub_technique_number': sub_technique_number,
                        'sub_technique_description': sub_technique_description
                    })
                    
df_subtechnique = pd.DataFrame(records)

df_subtechnique

tactics_list = []
for tactic in data:
    tactic_name = tactic.get('tactic-name')
    tactic_id = tactic.get('ID')
    tactic_description = tactic.get('tactic')
    tactics_list.append({'tactic_name': tactic_name, 'tactic_id': tactic_id, 'tactic_description': tactic_description})

df_tactic = pd.DataFrame(tactics_list)

df_tactic

technique_list = []
for tactic in data:
    tactic_name = tactic.get('tactic-name')
    tactic_id = tactic.get('ID')
    tactic_description = tactic.get('tactic')
    
    for technique_group in tactic.get('technique', []):
        for technique in technique_group:
            technique_name = technique.get('technique-name')
            technique_id = technique.get('technique-ID')
            technique_description = technique.get('technique-description')
            technique_list.append({
                'tactic_id': tactic_id,
                'technique_name': technique_name,
                'technique_id': technique_id,
                'technique_description': technique_description
            })

df_technique = pd.DataFrame(technique_list)

df_technique

def concatenate_strings(string_list):
    return '\n'.join(string_list)

df_subtechnique['sub_technique_description'] = df_subtechnique['sub_technique_description'].apply(concatenate_strings)
df_tactic['tactic_description'] = df_tactic['tactic_description'].apply(concatenate_strings)
df_technique['technique_description'] = df_technique['technique_description'].apply(concatenate_strings)

df_subtechnique

def merge_technique_and_sub(technique_id, sub_technique_number):
    return f"{technique_id}.{sub_technique_number}"

df_subtechnique['sub_technique_number'] = df_subtechnique.apply(lambda row: merge_technique_and_sub(row['technique_id'], row['sub_technique_number']), axis=1)

df_subtechnique

subtech_with_tech = df_subtechnique[['technique_id', 'sub_technique_number']]
subtech_with_tech

subtech_with_tech.to_csv(r'G:\NLP Projects\MitRA aTTACK\Notre projet\Data\Processed datasets\Matching\subtech_with_tech.csv', index=False)

subtech_names = df_subtechnique[['sub_technique_name', 'sub_technique_number']]
subtech_names

subtech_names.to_csv(r'G:\NLP Projects\MitRA aTTACK\Notre projet\Data\Processed datasets\Names\subtech_names.csv', index=False)

tech_with_tac = df_technique[['tactic_id', 'technique_id']]
tech_names = df_technique[['technique_name', 'technique_id']]

tech_with_tac.to_csv(r'G:\NLP Projects\MitRA aTTACK\Notre projet\Data\Processed datasets\Matching\tech_with_tac.csv', index=False)
tech_names.to_csv(r'G:\NLP Projects\MitRA aTTACK\Notre projet\Data\Processed datasets\Names\tech_names.csv', index=False)

tac_names = df_tactic[['tactic_name', 'tactic_id']]
tac_names.to_csv(r'G:\NLP Projects\MitRA aTTACK\Notre projet\Data\Processed datasets\Names\tac_names.csv', index=False)

df_tactic.to_csv(r'G:\NLP Projects\MitRA aTTACK\Notre projet\Data\Processed datasets\Description\tactic.csv', index=False)
df_technique.to_csv(r'G:\NLP Projects\MitRA aTTACK\Notre projet\Data\Processed datasets\Description\technique.csv', index=False)
df_subtechnique.to_csv(r'G:\NLP Projects\MitRA aTTACK\Notre projet\Data\Processed datasets\Description\subtechnique.csv', index=False)

df_tactic_2 = df_tactic.drop(columns=['tactic_name'])
df_tactic_2

df_technique_2 = df_technique.drop(columns=['tactic_id','technique_name'])
df_technique_2

df_subtechnique_2 = df_subtechnique.drop(columns=['technique_id','sub_technique_name'])
df_subtechnique_2

df_tactic_2.rename(columns={'tactic_id': 'id', 'tactic_description': 'desc'}, inplace=True)

df_technique_2.rename(columns={'technique_id': 'id', 'technique_description': 'desc'}, inplace=True)

df_subtechnique_2.rename(columns={'sub_technique_number': 'id', 'sub_technique_description': 'desc'}, inplace=True)

combined_df = pd.concat([df_tactic_2, df_technique_2, df_subtechnique_2])

combined_df.reset_index(drop=True, inplace=True)

combined_df

combined_df.to_csv(r'G:\NLP Projects\MitRA aTTACK\Notre projet\Data\Processed datasets\Final.csv', index=False)

df_subtechnique_3 = df_subtechnique[['technique_id','sub_technique_description']]
df_subtechnique_3

df_technique_3 = df_technique[['technique_id','technique_description']]
df_technique_3

df_technique_3.rename(columns={'technique_id': 'id', 'technique_description': 'desc'}, inplace=True)
df_subtechnique_3.rename(columns={'technique_id': 'id', 'sub_technique_description': 'desc'}, inplace=True)

combined_df_3 = pd.concat([df_technique_3, df_subtechnique_3])
combined_df_3.reset_index(drop=True, inplace=True)
combined_df_3

combined_df_3.to_csv(r'G:\NLP Projects\MitRA aTTACK\Notre projet\Data\Processed datasets\Final techniques.csv', index=False)
