==> Notebooks/5. SVC Predictions.ipynb <==
import pandas as pd
import pickle
import joblib

model = joblib.load(r'G:\NLP Projects\MitRA aTTACK\Notre projet\Models\ML Models\SVC.pkl')

vectorizer = joblib.load(r'G:\NLP Projects\MitRA aTTACK\Notre projet\Models\ML Models\vectorizer.joblib')

def pred(text):
    X = vectorizer.transform(text)
    predictions = model.predict(X)
    return predictions

Text = ["For websites and databases, the OWASP top 10 and CWE top 25 highlight the most common web-based vulnerabilities."]

print(pred(Text))

def pred(text):
    X = vectorizer.transform(text)
    predictions = model.predict(X)
    print(predictions)

Text = ["This account modification may immediately follow Create Account or other malicious account activity. Adversaries may also modify existing Valid Accounts that they have compromised."]

pred(Text)
